requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
cloudinary==1.23.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap4==2.3.1
django-heroku==0.3.1
django-registration==3.1.1
djangorestframework==3.12.2
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.4
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
